+++ app/controllers/posts_controller.rb
@@ -1,7 +1,7 @@
 class PostsController < ApplicationController
   def index
-    @posts = Post.all
     @user = User.find(params[:user_id])
+    @posts = @user.posts.includes(:comments)
   end
 
   def new
